refactor: bvh: fix runtime errors wrt memory and junk values
In places where pointers to pointers are used, the value need to be set to NULL for it to work properly. This is done by the "()" part of the "new[xyz]()". This essentially mimics calloc() for non user defined classes which includes
CustomClass **foo = new CustomClass* [num_elements]();
Here the array of pointers that is allocated is the non user defined class thus is set to 0 (NULL) using the []();